model/pyproject.toml
[project]
name = "image-text-ranking-model"
version = "0.1.0"
description = "Baseline model for image-text ranking system using dual-encoder architecture"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch~=2.7.0",
    "torchvision~=0.22.0",
    "transformers~=4.52.3",
    "pillow~=11.2.1",
    "numpy~=1.26.4",
    "scikit-learn~=1.5.0",
    "pandas~=2.2.3",
    "tqdm~=4.66.5",
    "matplotlib~=3.10.3",
    "requests~=2.32.3",
    "python-dotenv~=1.0.1",
    "mlflow~=2.21.0",
    "datasets~=3.6.0",
]

[tool.uv]
dev-dependencies = [
    "pytest~=8.3.5",
    "black~=24.3.0",
    "isort~=5.13.0",
    "flake8~=7.0.0",
    "jupyter~=1.0.0",
]
